Simpler install: one-command `uv sync` + `v2ecoli-parca` console script
Install before:
  clone the four sibling repos in a specific layout, uv sync, run a
  bash script to compile three Cython extensions, then python
  scripts/parca_run.py.

Install now:
  clone + `uv sync` + `v2ecoli-parca`.

Specifically:

- `[tool.uv.sources]` switched from local editable paths to git URLs
  on main / master for process-bigraph, bigraph-schema, and multi-cell.
  vEcoli now installs from PyPI (1.1.2.post3 — newer than the master
  tag and properly packages all sub-modules; master ships only the
  six top-level packages).
- `requires-python` bumped to ==3.12.12 to match vEcoli.
- `setup.py` (new) declares the three vendored Cython extensions as
  `ext_modules`, so `uv sync` compiles them automatically. The old
  `parca-build` extra is gone (cython + numpy are now build-system
  requirements). `scripts/parca_cython_build.sh` stays as a fallback
  for in-place rebuilds during .pyx development.
- `v2ecoli/cli/parca.py` (new) holds the CLI; registered as the
  `v2ecoli-parca` console script. Default `--cpus` is `os.cpu_count()`.
  `scripts/parca_run.py` becomes a thin shim so existing docs / tooling
  keep working.
- README rewrite: new top-level `## Install` section; the
  `### Re-running from scratch` block drops the manual Cython step.
- Allowlist `cli/parca.py` in `tests/test_pickle_allowlist.py` (the
  CLI writes the ParCa state pickle + step checkpoints — same
  ParCa-cache exception as `composite.py`).
